Moved files around for cleanliness and refactors to build properly.
