+++ SandBox/Makefile
@@ -1,0 +1,151 @@
+# Alternative GNU Make project makefile autogenerated by Premake
+
+ifndef config
+  config=debug
+endif
+
+ifndef verbose
+  SILENT = @
+endif
+
+.PHONY: clean prebuild
+
+SHELLTYPE := posix
+ifeq (.exe,$(findstring .exe,$(ComSpec)))
+	SHELLTYPE := msdos
+endif
+
+# Configurations
+# #############################################
+
+RESCOMP = windres
+INCLUDES += -I../Zital/Vendor/spdlog/include -I../Zital/Source -I../Zital/Vendor -I../Zital/Vendor/glm
+FORCE_INCLUDE +=
+ALL_CPPFLAGS += $(CPPFLAGS) -MD -MP $(DEFINES) $(INCLUDES)
+ALL_RESFLAGS += $(RESFLAGS) $(DEFINES) $(INCLUDES)
+LINKCMD = $(CXX) -o "$@" $(OBJECTS) $(RESOURCES) $(ALL_LDFLAGS) $(LIBS)
+define PREBUILDCMDS
+endef
+define PRELINKCMDS
+endef
+define POSTBUILDCMDS
+endef
+
+ifeq ($(config),debug)
+TARGETDIR = ../bin/Debug-windows-x86_64/Sandbox
+TARGET = $(TARGETDIR)/Sandbox.exe
+OBJDIR = ../bin-int/Debug-windows-x86_64/Sandbox
+DEFINES += -DZT_PLATFORM_WINDOWS -DZT_DEBUG
+ALL_CFLAGS += $(CFLAGS) $(ALL_CPPFLAGS) -m64 -g
+ALL_CXXFLAGS += $(CXXFLAGS) $(ALL_CPPFLAGS) -m64 -g -std=c++17
+LIBS += ../bin/Debug-windows-x86_64/Zital/Zital.lib
+LDDEPS += ../bin/Debug-windows-x86_64/Zital/Zital.lib
+ALL_LDFLAGS += $(LDFLAGS) -L/usr/lib64 -m64
+
+else ifeq ($(config),release)
+TARGETDIR = ../bin/Release-windows-x86_64/Sandbox
+TARGET = $(TARGETDIR)/Sandbox.exe
+OBJDIR = ../bin-int/Release-windows-x86_64/Sandbox
+DEFINES += -DZT_PLATFORM_WINDOWS -DZT_RELEASE
+ALL_CFLAGS += $(CFLAGS) $(ALL_CPPFLAGS) -m64 -O2
+ALL_CXXFLAGS += $(CXXFLAGS) $(ALL_CPPFLAGS) -m64 -O2 -std=c++17
+LIBS += ../bin/Release-windows-x86_64/Zital/Zital.lib
+LDDEPS += ../bin/Release-windows-x86_64/Zital/Zital.lib
+ALL_LDFLAGS += $(LDFLAGS) -L/usr/lib64 -m64 -s
+
+else ifeq ($(config),dist)
+TARGETDIR = ../bin/Dist-windows-x86_64/Sandbox
+TARGET = $(TARGETDIR)/Sandbox.exe
+OBJDIR = ../bin-int/Dist-windows-x86_64/Sandbox
+DEFINES += -DZT_PLATFORM_WINDOWS -DZT_DIST
+ALL_CFLAGS += $(CFLAGS) $(ALL_CPPFLAGS) -m64 -O2
+ALL_CXXFLAGS += $(CXXFLAGS) $(ALL_CPPFLAGS) -m64 -O2 -std=c++17
+LIBS += ../bin/Dist-windows-x86_64/Zital/Zital.lib
+LDDEPS += ../bin/Dist-windows-x86_64/Zital/Zital.lib
+ALL_LDFLAGS += $(LDFLAGS) -L/usr/lib64 -m64 -s
+
+endif
+
+# Per File Configurations
+# #############################################
+
+
+# File sets
+# #############################################
+
+GENERATED :=
+OBJECTS :=
+
+GENERATED += $(OBJDIR)/SandboxApp.o
+OBJECTS += $(OBJDIR)/SandboxApp.o
+
+# Rules
+# #############################################
+
+all: $(TARGET)
+	@:
+
+$(TARGET): $(GENERATED) $(OBJECTS) $(LDDEPS) | $(TARGETDIR)
+	$(PRELINKCMDS)
+	@echo Linking Sandbox
+	$(SILENT) $(LINKCMD)
+	$(POSTBUILDCMDS)
+
+$(TARGETDIR):
+	@echo Creating $(TARGETDIR)
+ifeq (posix,$(SHELLTYPE))
+	$(SILENT) mkdir -p $(TARGETDIR)
+else
+	$(SILENT) mkdir $(subst /,\\,$(TARGETDIR))
+endif
+
+$(OBJDIR):
+	@echo Creating $(OBJDIR)
+ifeq (posix,$(SHELLTYPE))
+	$(SILENT) mkdir -p $(OBJDIR)
+else
+	$(SILENT) mkdir $(subst /,\\,$(OBJDIR))
+endif
+
+clean:
+	@echo Cleaning Sandbox
+ifeq (posix,$(SHELLTYPE))
+	$(SILENT) rm -f  $(TARGET)
+	$(SILENT) rm -rf $(GENERATED)
+	$(SILENT) rm -rf $(OBJDIR)
+else
+	$(SILENT) if exist $(subst /,\\,$(TARGET)) del $(subst /,\\,$(TARGET))
+	$(SILENT) if exist $(subst /,\\,$(GENERATED)) del /s /q $(subst /,\\,$(GENERATED))
+	$(SILENT) if exist $(subst /,\\,$(OBJDIR)) rmdir /s /q $(subst /,\\,$(OBJDIR))
+endif
+
+prebuild: | $(OBJDIR)
+	$(PREBUILDCMDS)
+
+ifneq (,$(PCH))
+$(OBJECTS): $(GCH) | $(PCH_PLACEHOLDER)
+$(GCH): $(PCH) | prebuild
+	@echo $(notdir $<)
+	$(SILENT) $(CXX) -x c++-header $(ALL_CXXFLAGS) -o "$@" -MF "$(@:%.gch=%.d)" -c "$<"
+$(PCH_PLACEHOLDER): $(GCH) | $(OBJDIR)
+ifeq (posix,$(SHELLTYPE))
+	$(SILENT) touch "$@"
+else
+	$(SILENT) echo $null >> "$@"
+endif
+else
+$(OBJECTS): | prebuild
+endif
+
+
+# File Rules
+# #############################################
+
+$(OBJDIR)/SandboxApp.o: Source/SandboxApp.cpp
+	@echo "$(notdir $<)"
+	$(SILENT) $(CXX) $(ALL_CXXFLAGS) $(FORCE_INCLUDE) -o "$@" -MF "$(@:%.o=%.d)" -c "$<"
+
+-include $(OBJECTS:%.o=%.d)
+ifneq (,$(PCH))
+  -include $(PCH_PLACEHOLDER).d
+endif
