+++ Scripts/Win-GenerateProjects.bat
@@ -1,2 +1,6 @@
+@echo off
+
+pushd ..\
 call vendor\bin\premake\premake5.exe vs2022
+popd
 PAUSE
