+++ Zital/Source/Platform/Windows/WindowsWindow.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Zital/Window.h"
+#include "Zital/Core/Window.h"
 #include "Zital/Renderer/GraphicsContext.h"
 
 #include <GLFW/glfw3.h>
