+++ Zital/Source/Zital/Core/Log.cpp
