+++ Zital/Source/Zital/Core/Log.h
