+++ Zital/Source/Zital/ImGui/ImGuiLayer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Zital/Layer.h"
+#include "Zital/Core/Layer.h"
 #include "Zital/Events/KeyEvent.h"
 #include "Zital/Events/MouseEvent.h"
 #include "Zital/Events/ApplicationEvent.h"
