+++ Zital/Zital.vcxproj
@@ -151,24 +151,25 @@
     <ClInclude Include="Source\Platform\OpenGL\OpenGLVertexArray.h" />
     <ClInclude Include="Source\Platform\Windows\WindowsInput.h" />
     <ClInclude Include="Source\Platform\Windows\WindowsWindow.h" />
-    <ClInclude Include="Source\Zital\OrthographicCameraController.h" />
     <ClInclude Include="Source\ZTpch.h" />
     <ClInclude Include="Source\Zital.h" />
-    <ClInclude Include="Source\Zital\Application.h" />
-    <ClInclude Include="Source\Zital\Core.h" />
+    <ClInclude Include="Source\Zital\Core\Application.h" />
+    <ClInclude Include="Source\Zital\Core\Core.h" />
+    <ClInclude Include="Source\Zital\Core\EntryPoint.h" />
+    <ClInclude Include="Source\Zital\Core\Input.h" />
+    <ClInclude Include="Source\Zital\Core\KeyCodes.h" />
+    <ClInclude Include="Source\Zital\Core\Layer.h" />
+    <ClInclude Include="Source\Zital\Core\LayerStack.h" />
+    <ClInclude Include="Source\Zital\Core\Log.h" />
+    <ClInclude Include="Source\Zital\Core\MouseButtonCodes.h" />
     <ClInclude Include="Source\Zital\Core\Timestep.h" />
-    <ClInclude Include="Source\Zital\EntryPoint.h" />
+    <ClInclude Include="Source\Zital\Core\Window.h" />
     <ClInclude Include="Source\Zital\Events\ApplicationEvent.h" />
     <ClInclude Include="Source\Zital\Events\Event.h" />
     <ClInclude Include="Source\Zital\Events\KeyEvent.h" />
     <ClInclude Include="Source\Zital\Events\MouseEvent.h" />
     <ClInclude Include="Source\Zital\ImGui\ImGuiLayer.h" />
-    <ClInclude Include="Source\Zital\Input.h" />
-    <ClInclude Include="Source\Zital\KeyCodes.h" />
-    <ClInclude Include="Source\Zital\Layer.h" />
-    <ClInclude Include="Source\Zital\LayerStack.h" />
-    <ClInclude Include="Source\Zital\Log.h" />
-    <ClInclude Include="Source\Zital\MouseButtonCodes.h" />
+    <ClInclude Include="Source\Zital\OrthographicCameraController.h" />
     <ClInclude Include="Source\Zital\Renderer\Buffer.h" />
     <ClInclude Include="Source\Zital\Renderer\GraphicsContext.h" />
     <ClInclude Include="Source\Zital\Renderer\OrthographicCamera.h" />
@@ -178,7 +179,6 @@
     <ClInclude Include="Source\Zital\Renderer\Shader.h" />
     <ClInclude Include="Source\Zital\Renderer\Texture.h" />
     <ClInclude Include="Source\Zital\Renderer\VertexArray.h" />
-    <ClInclude Include="Source\Zital\Window.h" />
     <ClInclude Include="Vendor\glm\glm\common.hpp" />
     <ClInclude Include="Vendor\glm\glm\detail\_features.hpp" />
     <ClInclude Include="Vendor\glm\glm\detail\_fixes.hpp" />
@@ -606,16 +606,16 @@
     <ClCompile Include="Source\Platform\OpenGL\OpenGLVertexArray.cpp" />
     <ClCompile Include="Source\Platform\Windows\WindowsInput.cpp" />
     <ClCompile Include="Source\Platform\Windows\WindowsWindow.cpp" />
-    <ClCompile Include="Source\Zital\OrthographicCameraController.cpp" />
     <ClCompile Include="Source\ZTpch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="Source\Zital\Application.cpp" />
+    <ClCompile Include="Source\Zital\Core\Application.cpp" />
+    <ClCompile Include="Source\Zital\Core\Layer.cpp" />
+    <ClCompile Include="Source\Zital\Core\LayerStack.cpp" />
+    <ClCompile Include="Source\Zital\Core\Log.cpp" />
     <ClCompile Include="Source\Zital\ImGui\ImGuiBuild.cpp" />
     <ClCompile Include="Source\Zital\ImGui\ImGuiLayer.cpp" />
-    <ClCompile Include="Source\Zital\Layer.cpp" />
-    <ClCompile Include="Source\Zital\LayerStack.cpp" />
-    <ClCompile Include="Source\Zital\Log.cpp" />
+    <ClCompile Include="Source\Zital\OrthographicCameraController.cpp" />
     <ClCompile Include="Source\Zital\Renderer\Buffer.cpp" />
     <ClCompile Include="Source\Zital\Renderer\OrthographicCamera.cpp" />
     <ClCompile Include="Source\Zital\Renderer\RenderCommand.cpp" />
